Replace test utils with next module
