+++ aquadoggo/src/db/stores/task.rs
@@ -129,7 +129,7 @@ mod tests {
     use rstest::rstest;
 
     use crate::materializer::{Task, TaskInput};
-    use crate::test_utils::next::{test_runner, TestNode};
+    use crate::test_utils::{test_runner, TestNode};
 
     #[rstest]
     fn insert_get_remove_tasks(document_view_id: DocumentViewId) {
