+++ aquadoggo/src/graphql/client/dynamic_query.rs
@@ -427,7 +427,7 @@ mod test {
     use serde_json::json;
     use serial_test::serial;
 
-    use crate::test_utils::next::{
+    use crate::test_utils::{
         add_document, add_schema, graphql_test_client, test_runner, TestNode,
     };
