+++ aquadoggo/src/graphql/client/tests.rs
@@ -11,7 +11,7 @@ use rstest::rstest;
 use serde_json::json;
 use serial_test::serial;
 
-use crate::test_utils::next::{
+use crate::test_utils::{
     add_document, add_schema, graphql_test_client, populate_store_config, test_runner, TestNode,
 };
