+++ aquadoggo/src/http/service.rs
@@ -81,7 +81,7 @@ mod tests {
     use crate::graphql::GraphQLSchemaManager;
     use crate::http::context::HttpServiceContext;
     use crate::schema::SchemaProvider;
-    use crate::test_utils::next::{test_runner, TestNode};
+    use crate::test_utils::{test_runner, TestNode};
     use crate::test_utils::TestClient;
 
     use super::build_server;
