+++ aquadoggo/src/materializer/tasks/reduce.rs
@@ -235,7 +235,7 @@ mod tests {
 
     use crate::materializer::tasks::reduce_task;
     use crate::materializer::TaskInput;
-    use crate::test_utils::next::{
+    use crate::test_utils::{
         doggo_fields, doggo_schema, populate_store_config, test_runner, TestNode,
     };
