+++ aquadoggo/src/replication/service.rs
@@ -263,7 +263,7 @@ mod tests {
     use crate::replication::ReplicationConfiguration;
     use crate::test_utils::shutdown_handle;
     use crate::test_utils::{
-        doggo_fields, doggo_schema, with_db_manager_teardown, TestDatabaseManager,
+        doggo_fields, doggo_schema, test_runner_with_manager, TestNodeManager,
     };
     use crate::Configuration;
 
@@ -276,16 +276,12 @@ mod tests {
     }
     #[rstest]
     fn full_replication() {
-        with_db_manager_teardown(|db_manager: TestDatabaseManager| async move {
+        test_runner_with_manager(|manager: TestNodeManager| async move {
             init_env_logger();
 
             // Build and populate Billie's database
-            let billie_db = db_manager.create("sqlite::memory:").await;
-            billie_db
-                .context
-                .schema_provider
-                .update(doggo_schema())
-                .await;
+            let billie = manager.create("sqlite::memory:").await;
+            billie.context.schema_provider.update(doggo_schema()).await;
             let populate_db_config = PopulateStoreConfig {
                 no_of_entries: 1,
                 no_of_logs: 1,
@@ -295,19 +291,19 @@ mod tests {
                 create_operation_fields: doggo_fields(),
                 update_operation_fields: doggo_fields(),
             };
-            let (key_pairs, _) = populate_store(&billie_db.store, &populate_db_config).await;
+            let (key_pairs, _) = populate_store(&billie.store, &populate_db_config).await;
 
             // Launch HTTP service of Billie
             let (tx, _rx) = broadcast::channel(120);
             let tx_billie = tx.clone();
             let shutdown_billie = shutdown_handle();
             let context_billie = Context::new(
-                billie_db.store.clone(),
+                billie.context.store.clone(),
                 Configuration {
                     http_port: 3022,
                     ..Configuration::default()
                 },
-                billie_db.context.schema_provider.clone(),
+                billie.context.schema_provider.clone(),
             );
             let (tx_ready, rx_ready) = oneshot::channel::<()>();
 
@@ -338,7 +334,7 @@ mod tests {
                 },
                 ..Configuration::default()
             };
-            let ada_db = db_manager.create("sqlite::memory:").await;
+            let ada_db = manager.create("sqlite::memory:").await;
             ada_db.context.schema_provider.update(doggo_schema()).await;
             let context_ada = Context::new(
                 ada_db.store.clone(),
