+++ aquadoggo/src/test_utils/mod.rs
@@ -4,11 +4,14 @@ mod client;
 mod config;
 mod db;
 mod helpers;
+mod node;
 mod runner;
-pub mod next;
 
 pub use client::{graphql_test_client, shutdown_handle, TestClient};
 pub use config::{TestConfiguration, TEST_CONFIG};
-pub use db::{drop_database, initialize_db, initialize_db_with_url, TestData, TestDatabase};
-pub use helpers::{add_document, add_schema, build_document, doggo_fields, doggo_schema};
-pub use runner::{test_db, with_db_manager_teardown, TestDatabaseManager, TestDatabaseRunner};
+pub use db::{drop_database, initialize_db, initialize_db_with_url};
+pub use helpers::{build_document, doggo_fields, doggo_schema, schema_from_fields};
+pub use node::{
+    add_document, add_schema, populate_and_materialize, populate_store_config, TestNode,
+};
+pub use runner::{test_runner, test_runner_with_manager, TestNodeManager};
