+++ aquadoggo/src/test_utils/runner.rs
@@ -4,200 +4,132 @@ use std::panic;
 use std::sync::Arc;
 
 use futures::Future;
-use p2panda_rs::operation::OperationValue;
-use p2panda_rs::schema::Schema;
-use p2panda_rs::test_utils::memory_store::helpers::{populate_store, PopulateStoreConfig};
-use rstest::fixture;
 use tokio::runtime::Builder;
 use tokio::sync::Mutex;
 
-use crate::config::Configuration;
 use crate::context::Context;
 use crate::db::Pool;
 use crate::db::SqlStore;
 use crate::schema::SchemaProvider;
+use crate::test_utils::TestNode;
 use crate::test_utils::{initialize_db, initialize_db_with_url};
-use crate::test_utils::{TestData, TestDatabase};
-
-use super::{doggo_fields, doggo_schema};
+use crate::Configuration;
 
 #[async_trait::async_trait]
 pub trait AsyncTestFn {
-    async fn call(self, db: TestDatabase);
+    async fn call(self, db: TestNode);
 }
 
 #[async_trait::async_trait]
 impl<FN, F> AsyncTestFn for FN
 where
-    FN: FnOnce(TestDatabase) -> F + Sync + Send,
+    FN: FnOnce(TestNode) -> F + Sync + Send,
     F: Future<Output = ()> + Send,
 {
-    async fn call(self, db: TestDatabase) {
-        self(db).await
+    async fn call(self, node: TestNode) {
+        self(node).await
     }
 }
 
 #[async_trait::async_trait]
 pub trait AsyncTestFnWithManager {
-    async fn call(self, db: TestDatabaseManager);
+    async fn call(self, manager: TestNodeManager);
 }
 
 #[async_trait::async_trait]
 impl<FN, F> AsyncTestFnWithManager for FN
 where
-    FN: FnOnce(TestDatabaseManager) -> F + Sync + Send,
+    FN: FnOnce(TestNodeManager) -> F + Sync + Send,
     F: Future<Output = ()> + Send,
 {
-    async fn call(self, db: TestDatabaseManager) {
-        self(db).await
-    }
-}
-
-// @TODO: I'm keeping this here for now as otherwise we would need to refactor _all_ the tests
-// using it.
-//
-// We may still want to keep this "single database" runner injected through `rstest` but in any
-// case probably best to consider that in a different PR.
-pub struct TestDatabaseRunner {
-    config: PopulateStoreConfig,
-}
-
-impl TestDatabaseRunner {
-    /// Provides a safe way to write tests using a database which closes the pool connection
-    /// automatically when the test succeeds or fails.
-    ///
-    /// Takes an (async) test function as an argument and passes over the `TestDatabase` instance
-    /// so it can be used inside of it.
-    pub fn with_db_teardown<F: AsyncTestFn + Send + Sync + 'static>(&self, test: F) {
-        let runtime = Builder::new_current_thread()
-            .worker_threads(1)
-            .enable_all()
-            .thread_name("with_db_teardown")
-            .build()
-            .expect("Could not build tokio Runtime for test");
-
-        runtime.block_on(async {
-            // Initialise store
-            let pool = initialize_db().await;
-            let store = SqlStore::new(pool);
-
-            // Populate the store and construct test data
-            let (key_pairs, documents) = populate_store(&store, &self.config).await;
-            let test_data = TestData {
-                key_pairs,
-                documents,
-            };
-
-            // Init schema provider and insert the configured schema
-            let schema_provider = SchemaProvider::default();
-            schema_provider.update(self.config.schema.clone()).await;
-
-            // Construct the context
-            let context = Context::new(store.clone(), Configuration::default(), schema_provider);
-
-            // Construct the actual test database
-            let db = TestDatabase {
-                context,
-                store,
-                test_data,
-            };
-
-            // Get a handle of the underlying database connection pool
-            let pool = db.store.pool.clone();
-
-            // Spawn the test in a separate task to make sure we have control over the possible
-            // panics which might happen inside of it
-            let handle = tokio::task::spawn(async move {
-                // Execute the actual test
-                test.call(db).await;
-            });
-
-            // Get a handle of the task so we can use it later
-            let result = handle.await;
-
-            // Unwind the test by closing down the connection to the database pool. This will
-            // be reached even when the test panicked
-            pool.close().await;
-
-            // Panic here when test failed. The test fails within its own async task and stays
-            // there, we need to propagate it further to inform the test runtime about the result
-            match result {
-                Ok(_) => (),
-                Err(err) => panic::resume_unwind(err.into_panic()),
-            };
-        });
+    async fn call(self, manager: TestNodeManager) {
+        self(manager).await
     }
 }
 
 /// A manager which can create many databases and retain a handle on their connection pools.
 #[derive(Default)]
-pub struct TestDatabaseManager {
+pub struct TestNodeManager {
     pools: Arc<Mutex<Vec<Pool>>>,
 }
 
-impl TestDatabaseManager {
+impl TestNodeManager {
     pub fn new() -> Self {
         Self::default()
     }
 
-    pub async fn create(&self, url: &str) -> TestDatabase {
+    pub async fn create(&self, url: &str) -> TestNode {
         let pool = initialize_db_with_url(url).await;
 
         // Initialise test store using pool.
         let store = SqlStore::new(pool.clone());
 
-        let test_db = TestDatabase::new(store.clone());
+        let test_node = TestNode::new(store.clone());
 
         self.pools.lock().await.push(pool);
-        test_db
+        test_node
     }
 }
 
-/// Fixture for constructing a storage provider instance backed by a pre-populated database.
+/// Provides a safe way to write tests using a database which closes the pool connection
+/// automatically when the test succeeds or fails.
 ///
-/// Returns a `TestDatabaseRunner` that bootstraps a safe async test environment connecting to a
-/// database. It makes sure the runner disconnects properly from the connection pool after the test
-/// succeeded or even failed.
-///
-/// Passed parameters define what the database should contain. The first entry in each log contains
-/// a valid CREATE operation following entries contain duplicate UPDATE operations. If the
-/// with_delete flag is set to true the last entry in all logs contain be a DELETE operation.
-#[fixture]
-pub fn test_db(
-    // Number of entries per log/document
-    #[default(0)] no_of_entries: usize,
-    // Number of logs for each public key
-    #[default(0)] no_of_logs: usize,
-    // Number of authors, each with logs populated as defined above
-    #[default(0)] no_of_public_keys: usize,
-    // A boolean flag for wether all logs should contain a delete operation
-    #[default(false)] with_delete: bool,
-    // The schema used for all operations in the db
-    #[default(doggo_schema())] schema: Schema,
-    // The fields used for every CREATE operation
-    #[default(doggo_fields())] create_operation_fields: Vec<(&'static str, OperationValue)>,
-    // The fields used for every UPDATE operation
-    #[default(doggo_fields())] update_operation_fields: Vec<(&'static str, OperationValue)>,
-) -> TestDatabaseRunner {
-    let config = PopulateStoreConfig {
-        no_of_entries,
-        no_of_logs,
-        no_of_public_keys,
-        with_delete,
-        schema,
-        create_operation_fields,
-        update_operation_fields,
-    };
+/// Takes an (async) test function as an argument and passes over the `TestNode` instance
+/// so it can be used inside of it.
+pub fn test_runner<F: AsyncTestFn + Send + Sync + 'static>(test: F) {
+    let runtime = Builder::new_current_thread()
+        .worker_threads(1)
+        .enable_all()
+        .thread_name("with_db_teardown")
+        .build()
+        .expect("Could not build tokio Runtime for test");
 
-    TestDatabaseRunner { config }
+    runtime.block_on(async {
+        // Initialise store
+        let pool = initialize_db().await;
+        let store = SqlStore::new(pool);
+
+        // Construct the actual test node
+        let node = TestNode {
+            context: Context::new(
+                store.clone(),
+                Configuration::default(),
+                SchemaProvider::default(),
+            ),
+        };
+
+        // Get a handle of the underlying database connection pool
+        let pool = node.context.store.pool.clone();
+
+        // Spawn the test in a separate task to make sure we have control over the possible
+        // panics which might happen inside of it
+        let handle = tokio::task::spawn(async move {
+            // Execute the actual test
+            test.call(node).await;
+        });
+
+        // Get a handle of the task so we can use it later
+        let result = handle.await;
+
+        // Unwind the test by closing down the connection to the database pool. This will
+        // be reached even when the test panicked
+        pool.close().await;
+
+        // Panic here when test failed. The test fails within its own async task and stays
+        // there, we need to propagate it further to inform the test runtime about the result
+        match result {
+            Ok(_) => (),
+            Err(err) => panic::resume_unwind(err.into_panic()),
+        };
+    });
 }
 
 /// Method which provides a safe way to write tests with the ability to build many databases and
 /// have their pool connections closed automatically when the test succeeds or fails.
 ///
-/// Takes an (async) test function as an argument and passes over the `TestDatabaseManager`
+/// Takes an (async) test function as an argument and passes over the `TestNodeManager`
 /// instance which can be used to build databases from inside the tests.
-pub fn with_db_manager_teardown<F: AsyncTestFnWithManager + Send + Sync + 'static>(test: F) {
+pub fn test_runner_with_manager<F: AsyncTestFnWithManager + Send + Sync + 'static>(test: F) {
     let runtime = Builder::new_current_thread()
         .worker_threads(1)
         .enable_all()
@@ -206,17 +138,17 @@ pub fn with_db_manager_teardown<F: AsyncTestFnWithManager + Send + Sync + 'stati
         .expect("Could not build tokio Runtime for test");
 
     // Instantiate the database manager
-    let db_manager = TestDatabaseManager::new();
+    let manager = TestNodeManager::new();
 
     // Get a handle onto it's collection of pools
-    let pools = db_manager.pools.clone();
+    let pools = manager.pools.clone();
 
     runtime.block_on(async {
         // Spawn the test in a separate task to make sure we have control over the possible
         // panics which might happen inside of it
         let handle = tokio::task::spawn(async move {
             // Execute the actual test
-            test.call(db_manager).await;
+            test.call(manager).await;
         });
 
         // Get a handle of the task so we can use it later
